Major Fixes

+++ bfs_orig/cd_spnet.h
@@ -31,7 +31,7 @@ const int GA_DEBUG_L2 = 0; // level 2 debugging
 
 const double GA_RANDOM_POP_PERCENT = 0.25; // set a minimum of 20 percent here
 const int GA_TOURNAMENT_SIZE = 7; // tournament size for the seleciton operator
-const int GA_CROSSOVER_SIZE_RATE = 0.6; // percentage of chromosome used for crossover
+const double GA_CROSSOVER_SIZE_RATE = 0.6; // percentage of chromosome used for crossover
 const double GA_CROSSOVER_RATE = 0.85; // max chr size used for crossover
 const double GA_REPRODUCTION_RATE = 0.05; // rate of reproduction producing the offsprings
 const double GA_MUTATION_RATE = 0.25; // rate of mutation producing the offspring
